+++ src/ModuleServiceProvider.php
@@ -6,6 +6,9 @@ namespace SebastiaanLuca\Module;
 
 use Illuminate\Filesystem\Filesystem;
 use Illuminate\Support\ServiceProvider;
+use SebastiaanLuca\Module\Commands\Cache;
+use SebastiaanLuca\Module\Commands\ClearCache;
+use SebastiaanLuca\Module\Commands\CreateModule;
 use SebastiaanLuca\Module\Commands\RefreshModules;
 use SebastiaanLuca\Module\Commands\RegisterModuleAutoloading;
 use SebastiaanLuca\Module\Services\ModuleLoader;
@@ -22,15 +25,19 @@ class ModuleServiceProvider extends ServiceProvider
         $this->app->singleton(ModuleLoader::class, function () {
             return new ModuleLoader(
                 app(Filesystem::class),
-                config($this->getShortPackageName())
+                config($this->getShortPackageName()),
+                require base_path('vendor/autoload.php')
             );
         });
 
         $this->app->bind($this->getShortPackageName(), ModuleLoader::class);
 
         $this->commands([
+            CreateModule::class,
             RegisterModuleAutoloading::class,
             RefreshModules::class,
+            Cache::class,
+            ClearCache::class,
         ]);
 
         app(ModuleLoader::class)->load(