+++ tests/resources/cache.php
@@ -1,0 +1,3 @@
+<?php return array (
+    1 => 'MyModule\\Providers\\MyModuleServiceProvider',
+);