+++ tests/resources/setup/modules/Another/src/Providers/AnotherServiceProvider.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Another\Providers;
+
+use SebastiaanLuca\Module\Providers\ModuleProvider;
+
+class AnotherServiceProvider extends ModuleProvider
+{
+}