+++ tests/resources/setup/modules/MyModule/src/Providers/MyModuleServiceProvider.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MyModule\Providers;
+
+use SebastiaanLuca\Module\Providers\ModuleProvider;
+
+class MyModuleServiceProvider extends ModuleProvider
+{
+}